requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.4.0
Django==4.1.7
django-celery-beat==2.5.0
django-filter==23.2
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
kombu==5.2.4
Markdown==3.4.3
prompt-toolkit==3.0.38
psycopg2==2.9.5
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
validate-email==1.3
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
stripe==5.4.0
python-dotenv==1.0.0